+++ pkg/mcast/network/reliable_transport.go
@@ -1,4 +1,4 @@
-package core
+package network
 
 import (
 	"context"
